Server/requirements.txt
pyppeteer==1.0.2
beautifulsoup4
flask
flask-cors
schedule
python-dotenv
requests